refactored render stuff to render/ dir
