+++ vs/stdafx.h
@@ -9,9 +9,9 @@
 #include "debug_input.h"
 //#include "events.h"
 #include "physics.h"
-#include "render.h"
+#include "render/render.h"
 #include "scene_manager.h"
-#include "shadows.h"
+#include "render/shadows.h"
 #include "timer.h"
 #include "components/transform.h"
 #include "window.h"
